+++ CHANGELOG.md
@@ -37,6 +37,8 @@
     not match 'TeamMembers' parameter - fixes [Issue #342](https://github.com/PowerShell/NetworkingDsc/issues/342).
 - Updated examples to format required for publishing to PowerShell Gallery - fixes
   [Issue #374](https://github.com/PowerShell/NetworkingDsc/issues/374).
+- New Resource: NetAdapterState to enable or disable a network adapter - fixes
+  [Issue #365](https://github.com/PowerShell/NetworkingDsc/issues/365)
 
 ## 6.3.0.0